+++ api/db/migrations/20220508212212_add_currency_to_rate/migration.sql
@@ -1,0 +1,19 @@
+-- RedefineTables
+PRAGMA foreign_keys=OFF;
+CREATE TABLE "new_Rate" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "value" REAL NOT NULL,
+    "currency" TEXT NOT NULL DEFAULT 'USD',
+    "name" TEXT NOT NULL,
+    "type" TEXT NOT NULL,
+    "material" TEXT NOT NULL,
+    "unit" TEXT NOT NULL,
+    "description" TEXT,
+    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" DATETIME NOT NULL
+);
+INSERT INTO "new_Rate" ("createdAt", "description", "id", "material", "name", "type", "unit", "updatedAt", "value") SELECT "createdAt", "description", "id", "material", "name", "type", "unit", "updatedAt", "value" FROM "Rate";
+DROP TABLE "Rate";
+ALTER TABLE "new_Rate" RENAME TO "Rate";
+PRAGMA foreign_key_check;
+PRAGMA foreign_keys=ON;