+++ scripts/seed.js
@@ -8,18 +8,11 @@ export default async () => {
     //
     // Update "const data = []" to match your data model and seeding needs
     //
-    const data = [
-      // To try this example data with the UserExample model in schema.prisma,
-      // uncomment the lines below and run 'yarn rw prisma migrate dev'
-      //
-      // { name: 'alice', email: 'alice@example.com' },
-      // { name: 'mark', email: 'mark@example.com' },
-      // { name: 'jackie', email: 'jackie@example.com' },
-      // { name: 'bob', email: 'bob@example.com' },
-    ]
-    console.log(
-      "\nUsing the default './scripts/seed.{js,ts}' template\nEdit the file to add seed data\n"
-    )
+    const latestRates = require('./latest-rates.json')
+    const data = latestRates
+    // console.log(
+    //   "\nUsing the default './scripts/seed.{js,ts}' template\nEdit the file to add seed data\n"
+    // )
 
     // Note: if using PostgreSQL, using `createMany` to insert multiple records is much faster
     // @see: https://www.prisma.io/docs/reference/api-reference/prisma-client-reference#createmany
@@ -28,8 +21,9 @@ export default async () => {
       // Change to match your data model and seeding needs
       //
       data.map(async (data) => {
-        const record = await db.userExample.create({ data })
+        const record = await db.rate.create({ data })
         console.log(record)
+        console.info('++ ' + record.type + ' ' + record.material + ' ...')
       })
     )
   } catch (error) {