+++ web/src/components/FutureImperfectHeader.js
@@ -21,6 +21,11 @@ const FutureImperfectHeader = ({ children }) => {
               </NavLink>
             </li>
             <li>
+              <NavLink activeClassName="active" to={routes.rates()}>
+                Rates
+              </NavLink>
+            </li>
+            <li>
               <NavLink activeClassName="active" to={routes.chapter1()}>
                 Ch: 1
               </NavLink>