+++ web/src/components/PriceSheet/PriceSheet.stories.js
@@ -1,0 +1,7 @@
+import PriceSheet from './PriceSheet'
+
+export const generated = () => {
+  return <PriceSheet />
+}
+
+export default { title: 'Components/PriceSheet' }