+++ web/src/pages/Rate/EditRatePage/EditRatePage.js
@@ -1,0 +1,7 @@
+import EditRateCell from 'src/components/Rate/EditRateCell'
+
+const EditRatePage = ({ id }) => {
+  return <EditRateCell id={id} />
+}
+
+export default EditRatePage