+++ web/src/pages/Rate/RatePage/RatePage.js
@@ -1,0 +1,7 @@
+import RateCell from 'src/components/Rate/RateCell'
+
+const RatePage = ({ id }) => {
+  return <RateCell id={id} />
+}
+
+export default RatePage